creacion clases model
