--- src/main/java/liquidaciondiversos/Risk.java
@@ -1,9 +1,0 @@
-package liquidaciondiversos;
-
-import lombok.Data;
-
-@Data
-public class Risk {
-    private String code;
-    private String name;
-}
