+++ src/main/java/liquidaciondiversos/model/Client.java
@@ -1,0 +1,15 @@
+package liquidaciondiversos.model;
+
+import lombok.Data;
+
+@Data
+public class Client {
+
+    private String name;
+    private String surname;
+    private String direction;
+    private String telephone;
+    private String nif;
+    private Warranty warranty;
+
+}
