+++ src/main/java/liquidaciondiversos/model/Policy.java
@@ -1,0 +1,11 @@
+package liquidaciondiversos.model;
+
+import lombok.Data;
+
+@Data
+public class Policy {
+
+    private Product product;
+    private Client client;
+    private double capitals;
+}
