+++ src/main/java/liquidaciondiversos/model/Product.java
@@ -1,0 +1,15 @@
+package liquidaciondiversos.model;
+
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class Product {
+    private String code;
+    private String name;
+    List<ProductWarranty> productWarranties;
+
+
+}
