+++ src/main/java/liquidaciondiversos/model/Sinister.java
@@ -1,0 +1,14 @@
+package liquidaciondiversos.model;
+
+import lombok.Data;
+
+import java.util.Date;
+import java.util.List;
+
+@Data
+public class Sinister {
+    private Policy policy;
+    private Date date;
+    private String cause;
+    private List<ProductWarranty> damagedList;
+}
