+++ src/main/java/liquidaciondiversos/model/WarrantyType.java
@@ -1,4 +1,4 @@
-package liquidaciondiversos;
+package liquidaciondiversos.model;
 
 public enum WarrantyType {
     CONTENT,
